Move education group type autocomplete to education_group app
