+++ base/views/education_groups/urls.py
@@ -30,7 +30,6 @@ from base.views.education_groups.clear_clipboard import clear_clipboard
 from base.views.education_groups.publication_contact import CreateEducationGroupPublicationContactView, \
     UpdateEducationGroupPublicationContactView, EducationGroupPublicationContactDeleteView, \
     UpdateEducationGroupEntityPublicationContactView
-from base.views.education_groups.search import EducationGroupTypeAutoComplete
 from base.views.education_groups.update import CertificateAimAutocomplete
 from education_group import urls as education_group_urls
 from .achievement.urls import urlpatterns as urlpatterns_achievement
@@ -41,12 +40,6 @@ urlpatterns = [
         CertificateAimAutocomplete.as_view(),
         name='certificate_aim_autocomplete',
     ),
-    url(
-        r'^education_group_type_autocomplete/$',
-        EducationGroupTypeAutoComplete.as_view(),
-        name='education_group_type_autocomplete'
-    ),
-
     url(
         r'^clear_clipboard/$',
         clear_clipboard,
