+++ education_group/urls.py
@@ -9,6 +9,7 @@ from base.views.education_groups.achievement.update import EducationGroupAchieve
 from education_group.converters import GroupTypeConverter, TrainingTypeConverter, MiniTrainingTypeConverter, \
     AcronymConverter
 from education_group.views import group, training, mini_training, general_information
+from education_group.views.autocomplete import EducationGroupTypeAutoComplete
 from education_group.views.configuration.common_list import CommonListView
 from education_group.views.configuration.home import ConfigurationHomeView
 from education_group.views.mini_training.delete import MiniTrainingDeleteView
@@ -241,5 +242,12 @@ urlpatterns = [
     path('configuration/', include([
         path('home/', ConfigurationHomeView.as_view(), name='catalog_configuration'),
         path('common-topics/', CommonListView.as_view(), name='common_topics_configuration'),
-    ]))
+    ])),
+    path('autocomplete/', include([
+        path(
+            'education-group-types/',
+            EducationGroupTypeAutoComplete.as_view(),
+            name='education_group_type_autocomplete'
+        ),
+    ])),
 ]
